requirements.txt
cloudpickle==0.4.0
exifread==2.1.2
flask==2.3.2
fpdf2==2.4.6
joblib==0.14.1
matplotlib
networkx==2.5
numpy>=1.19
Pillow>=8.1.1
pyproj>=1.9.5.1
pytest==3.0.7
python-dateutil>=2.7
pyyaml==5.4
scipy>=1.10.0
Sphinx==4.2.0
six
xmltodict==0.10.2
wheel
opencv-python==4.5.1.48 ; sys_platform == "win32"
opencv-python ; sys_platform == "linux"
flask-sqlalchemy==3.1.1
greenlet==3.0.3
sqlalchemy==2.0.29
alembic==1.13.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
flask_cors==4.0.1
